epistemic/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.30.0
httpx==0.28.1
pydantic-settings==2.1.0
authlib==1.3.0
itsdangerous==2.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.25
aiosqlite==0.19.0
